+++ src/cli/index.ts
@@ -1,55 +1,136 @@
-import { Command, CommandArg, ToolController } from "../core";
-import { controllers } from "../controllers";
-import { consoleTerminal, formatTable } from "../core/utils";
+import {
+    Command,
+    CommandArg,
+    ToolController,
+} from "../core";
+import {controllers} from "../controllers";
+import {
+    consoleTerminal,
+    formatTable,
+} from "../core/utils";
 import * as fs from "fs";
 import * as path from "path";
 
-enum ParseState {
-    OptionOrArg,
-    OptionValue,
+function argIsOption(arg: CommandArg, name: string, alias: string): boolean {
+    return !arg.isArg && (
+        name !== "" && arg.name === name
+        || alias !== "" && arg.alias === alias
+    );
 }
 
-type CommandLine = {
-    args: string[];
-    options: { [name: string]: any };
+function findOptionArg(command: Command, name: string): CommandArg | undefined {
+    const full = name.startsWith("--") ? name.substr(2).split("-").join("") : "";
+    const alias = name.startsWith("-") ? name.substr(1) : "";
+    return command.args?.find(arg => argIsOption(arg, full, alias));
 }
 
-function parseCommandLine(programArgs: string[]): CommandLine {
-    let state = ParseState.OptionOrArg;
-    let option = "";
-    const args: string[] = [];
-    const options: { [name: string]: any } = {};
+class CommandLine {
+    args: { help?: boolean, [name: string]: any } = {};
+    controller: ToolController | undefined = undefined;
+    command: Command | undefined = undefined;
+    positional: CommandArg[] = [];
+    unresolved = new Map<string, CommandArg>();
+    pending: CommandArg | undefined = undefined;
 
-    for (const arg of programArgs) {
-        const argOption: string | null =
-            arg.startsWith("--") ? arg.substr(2) : (arg.startsWith("-") ? arg : null);
-        switch (state) {
-            case ParseState.OptionOrArg:
-                if (option !== "") {
-                    options[option] = "";
-                }
-                if (argOption) {
-                    option = argOption;
-                    state = ParseState.OptionValue;
-                } else {
-                    option = "";
-                    args.push(arg);
-                }
-                break;
-            case ParseState.OptionValue:
-                options[option] = arg;
-                option = "";
-                state = ParseState.OptionOrArg;
-                break;
+    resolveValue(arg: CommandArg, value: string | undefined) {
+        if (arg.type === "boolean" && value === undefined) {
+            value = "true";
+        }
+        if (value === undefined) {
+            throw new Error(`Missing value for ${arg.name}`);
+        }
+        let resolved: any;
+        if (arg.type === "boolean") {
+            resolved = value.toLowerCase() === "true";
+        } else {
+            resolved = value;
+        }
+        if (arg.getVariants !== undefined && !arg.getVariants().find(x => x.name === value)) {
+            throw missingArgError(arg);
+        }
+        this.args[arg.name] = resolved;
+        this.unresolved.delete(arg.name);
+        const i = this.positional.indexOf(arg);
+        if (i >= 0) {
+            this.positional.splice(i, 1);
+        }
+        if (this.pending === arg) {
+            this.pending = undefined;
         }
     }
-    if (option !== "") {
-        options[option] = "";
+
+    resolveDefault(arg: CommandArg) {
+        if (arg.defaultValue !== undefined) {
+            if (arg.type === "boolean") {
+                this.args[arg.name] = arg.defaultValue === "true";
+            } else {
+                this.args[arg.name] = arg.defaultValue;
+            }
+        } else if (arg.type === "folder") {
+            this.args[arg.name] = process.cwd();
+        } else {
+            throw missingArgError(arg);
+        }
     }
-    return {
-        args,
-        options,
-    };
+
+    parseOptionName(name: string) {
+        if (this.pending) {
+            this.resolveValue(this.pending, undefined);
+        }
+        const optionName: string = name.toLowerCase();
+        if (optionName === "--help" || optionName === "-h") {
+            this.args.help = true;
+        } else if (this.command) {
+            this.pending = findOptionArg(this.command, optionName);
+            if (!this.pending) {
+                throw new Error(`Unknown option ${optionName}`);
+            }
+            if (this.pending.type === "boolean") {
+                this.resolveValue(this.pending, undefined);
+            }
+        } else if (this.controller) {
+            throw new Error(`Unexpected option ${optionName} before command name.`);
+        } else {
+            throw new Error(`Unexpected option ${optionName} before tool name.`);
+        }
+    }
+
+    parse(programArgs: string[]) {
+        for (const arg of programArgs) {
+            if (arg.startsWith("-")) {
+                this.parseOptionName(arg);
+            } else if (this.pending) {
+                this.resolveValue(this.pending, arg);
+            } else if (this.controller && this.command) {
+                if (this.positional.length === 0) {
+                    throw new Error(`Unexpected argument ${arg}`);
+                }
+                this.resolveValue(this.positional[0], arg);
+            } else if (this.controller) {
+                this.command = this.controller.commands.find(x => x.name === arg.toLowerCase());
+                if (!this.command) {
+                    throw new Error(`Unknown command: ${arg}`);
+                }
+                for (const arg of this.command.args ?? []) {
+                    this.unresolved.set(arg.name, arg);
+                    if (arg.isArg) {
+                        this.positional.push(arg);
+                    }
+                }
+            } else {
+                this.controller = controllers.find(x => x.name === arg.toLowerCase());
+                if (!this.controller) {
+                    throw new Error(`Unknown tool: ${arg}.`);
+                }
+            }
+        }
+        if (this.pending) {
+            this.resolveValue(this.pending, undefined);
+        }
+        for (const arg of this.unresolved.values()) {
+            this.resolveDefault(arg);
+        }
+    }
 }
 
 export function printCommandUsage(controller: ToolController, command: Command) {
@@ -73,14 +154,16 @@ export function printCommandUsage(controller: ToolController, command: Command) 
         console.log(formatTable(args.map(x => ["  ", x.name, x.title])));
     }
     console.log("Options:");
-    console.log(`    --help, -h  Show command usage`);
-    console.log(formatTable(options.map(x => ["  ", `--${x.name}`, x.title])));
+    console.log(formatTable([
+        ["  ", "--help, -h", "Show command usage"],
+        ...options.map(x => ["  ", `--${x.name}${x.alias ? ", -" + x.alias : ""}`, x.title]),
+    ]));
     return;
 }
 
 export function printControllerUsage(controller: ToolController) {
     const commands: [string, Command][] = controller.commands
-        .map(x => [`${controller.name} ${x.name}`, x]);
+    .map(x => [`${controller.name} ${x.name}`, x]);
     console.log(formatTable(commands.map(x => ["  ", x[0], x[1].title])));
 }
 
@@ -108,78 +191,29 @@ export function printUsage(controller?: ToolController, command?: Command) {
 function missingArgError(arg: CommandArg): Error {
     const variants: string = arg.getVariants
         ? "\n" +
-          formatTable(
-              [["Available variants:", ""], ...arg.getVariants().map(x => [x.name, x.description ?? ""])],
-              { headerSeparator: true }
-          )
+        formatTable(
+            [["Available variants:", ""], ...arg.getVariants().map(x => [x.name, x.description ?? ""])],
+            {headerSeparator: true},
+        )
         : "";
     throw new Error(`Missing required ${arg.name}${variants}`);
 }
 
-function getArgValue(arg: CommandArg, commandLine: CommandLine): string | undefined {
-    if (arg.isArg) {
-        const value = commandLine.args.splice(0, 1)[0];
-        if (value !== undefined) {
-            if (arg.getVariants === undefined) return value;
-            if (arg.getVariants().find(x => x.name === value)) return value;
-            throw missingArgError(arg);
-        }
-        if (arg.defaultValue !== undefined) {
-            return arg.defaultValue;
-        }
-        throw missingArgError(arg);
-    }
-    let value = commandLine.options[arg.name];
-    if (value !== undefined) {
-        return value;
-    }
-    if (arg.defaultValue !== undefined) {
-        return arg.defaultValue;
-    }
-    if (arg.type === "folder") {
-        return process.cwd();
-    }
-    throw missingArgError(arg);
-}
-
-function extractNextArg(commandLine: CommandLine): string {
-    return (commandLine.args.splice(0, 1)[0] ?? "").toLowerCase();
-}
-
-function getOption(commandLine: CommandLine, name: string, alias: string): string | undefined {
-    if (commandLine.options[name] !== undefined) {
-        return commandLine.options[name];
-    }
-    return commandLine.options[alias];
-}
-
 export async function run() {
-    const commandLine = parseCommandLine(process.argv.slice(2));
-    if (commandLine.args.length === 0) {
-        printUsage();
-        return;
-    }
-    const controllerName = extractNextArg(commandLine);
-    const controller = controllers.find(x => x.name === controllerName);
-    if (!controller) {
-        throw new Error(`Unknown tool: ${controllerName}.`);
-    }
-    const commandName = extractNextArg(commandLine);
-    const command = controller.commands.find(x => x.name === commandName);
-    if (!command) {
-        if (commandName) {
-            throw new Error(`Unknown command: ${commandName}`);
-        }
+    const parser = new CommandLine();
+    parser.parse(process.argv.slice(2));
+    const {
+        controller,
+        command,
+        args,
+    } = parser;
+    if (!controller || !command) {
         printUsage(controller, command);
         return;
     }
-    if (getOption(commandLine, "help", "-h") !== undefined) {
+    if (parser.args.help) {
         printUsage(controller, command);
         return;
-    }
-    const args: { [name: string]: any } = {};
-    for (const arg of command.args ?? []) {
-        args[arg.name] = getArgValue(arg, commandLine);
     }
     await command.run(consoleTerminal, args);
 }